Remove the wass est step at the end of each loop
In the past run, over the course of 10 epochs the wasserstein
loss increased over the first couple hundered iterations and then
stayed at a constant value. In the training loop, at the end of
the loop there was a step where training was set to false in
the generator. I'm wondering if that setting somehow persists and
the generator never actually trains. Doing a run with that step
removed should answer this question.