Only train critic for 4 epochs
The last run still had the loss go up to a certain value and not
come down in 10 epochs. To determine whether this is an issue
with the way training is being done or is an actual bug in the
the way gradient descent is happening on the generator, I'll have
both nets trained for 4 epochs, then just the generator, which
should ensure that the loss will go down.