Get rid of batch_norm, switch to LeakyReLU
The gradients on the generator were going to zero when the activation
function was just ReLU - switching to LeakyReLU appeared to fix the
issue with the generator not training. I also got rid of the batch
norm layers in an attempt to simplify everything. It might be worth
experimenting with putting them back in later on.