Make concatfnx work with TF tensors, fix dtype bug
Due to how the concatenation function is used in the train script,
it needs to work with tf tensors instead of numpy arrays. Also
fixed bug with the load_mnist_data function returning tensors with
two different data types (float32 and float64), which was causing issues
with the concat function.