Change sampling method for GP distribution
The gradient norm is supposed to be enforced on "straight lines between
points in the real and generated distributions." Because I'm dealing
with a conditional distribution, I wasn't sure whether my straight
lines should be between points with the same x values or different
values. Previosly, I had been using points with the same x values.
This commit changes the interpolation function, as well as the
train_critic function in the cWGAN class and the take_critic_step
function in the Trainer class in order to use points with different
x values.