Add functionality for clustering parton jets
Change the program to use two pythia objects - one that processes the
hard scatter + parton shower, and then one that does hadronization. This
allows us to grab the partons at the end of the shower and cluster them into
jets, and then continue on with the simulation so we can use Delphes and get
the reco jets as well.